+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+# Feat
+
+- store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
+
 ### v7.0.2
 
 # Feat